requirements-fixer.txt
google-cloud-compute==1.46.0
google-api-core==2.30.0
google-cloud-resource-manager==1.16.0